PIE-250 - moved boosted modal into a component
